DRYed out some descritpive methods for factory/updater services
Whitelisted all descriptive attributes in digital_object/descriptive controllers [skip ci]
