+++ app/controllers/curator/digital_objects_controller.rb
@@ -44,7 +44,41 @@ module Curator
                                                is_member_of_collection: [:ark_id],
                                                contained_by: [:ark_id],
                                                metastreams: {
-                                                 descriptive: {},
+                                                 descriptive: [:abstract,
+                                                               :access_restrictions,
+                                                               :digital_origin,
+                                                               :extent,
+                                                               :frequency,
+                                                               :issuance,
+                                                               :origin_event,
+                                                               :physical_location_department,
+                                                               :physical_location_shelf_locator,
+                                                               :place_of_publication,
+                                                               :publisher,
+                                                               :resource_type_manuscript,
+                                                               :rights,
+                                                               :series,
+                                                               :subseries,
+                                                               :subsubseries,
+                                                               :text_direction,
+                                                               :toc,
+                                                               :toc_url,
+                                                               cartographic: [:projection, scale: []],
+                                                               date: [:created, :issued, :copyright],
+                                                               host_collections: [],
+                                                               identifier: [:label, :type, :invalid],
+                                                               genres: [:label, :id_from_auth, :authority_code, :basic],
+                                                               languages: [:label, :id_from_auth, :authority_code],
+                                                               license: [:label, :uri],
+                                                               name_roles: [name: [:label, :name_type, :authority_code, :id_from_auth, :affiliation], role: [:label, :id_from_auth, :authority_code]],
+                                                               note: [:label, :type],
+                                                               physical_location: [:authority_code, :id_from_auth, :label, :affiliation, :name_type],
+                                                               publication: [:edition_name, :edition_number, :volume, :issue_number],
+                                                               related: [:constituent, referenced_by_url: [], references_url: [], other_format: [], review_url: []],
+                                                               resource_types: [:label, :authority_code, :id_from_auth],
+                                                               subject: [topics: [:label, :authority_code, :id_from_auth], names: [:label, :name_type, :authority_code, :id_from_auth, :affiliation], geos: [:label, :authority_code, :id_from_auth, :coordinates, :bounding_box, :area_type], titles: [:label, :id_from_auth, :subtitle, :authority_code, :display, :display_label, :usage, :supplied, :language, :type, :part_number, :part_name], temporals: [], dates: []],
+                                                               title: [primary: [:label, :id_from_auth, :authority_code, :subtitle, :display, :display_label, :usage, :supplied, :language, :type, :part_number, :part_name], other: [:label, :id_from_auth, :subtitle, :authority_code, :display, :display_label, :usage, :supplied, :language, :type, :part_number, :part_name]]
+                                                             ],
                                                                administrative: [:description_standard, :hosting_status, :harvestable, :flagged, destination_site: [], access_edit_group: []],
                                                                workflow: [:ingest_origin, :publishing_state, :processing_state]
                                                })
