+++ spec/controllers/curator/digital_objects_controller_spec.rb
@@ -24,37 +24,4 @@ RSpec.describe Curator::DigitalObjectsController, type: :controller do
   let(:serializer_class) { Curator::DigitalObjectSerializer }
 
   include_examples 'shared_formats', include_ark_context: true, skip_post: false, resource_key: 'digital_object'
-
-  skip "PUT #update" do
-    context "with valid params" do
-      let(:new_attributes) {
-        skip("Add a hash of attributes valid for your model")
-      }
-
-      it "updates the requested digital_object" do
-        digital_object = Curator::DigitalObject.create! valid_attributes
-        put :update, params: { id: digital_object.to_param, digital_object: new_attributes }, session: valid_session
-        digital_object.reload
-        skip("Add assertions for updated state")
-      end
-
-      it "renders a JSON response with the digital_object" do
-        digital_object = Curator::DigitalObject.create! valid_attributes
-
-        put :update, params: { id: digital_object.to_param, digital_object: valid_attributes }, session: valid_session
-        expect(response).to have_http_status(:ok)
-        expect(response.content_type).to eq('application/json')
-      end
-    end
-
-    context "with invalid params" do
-      it "renders a JSON response with errors for the digital_object" do
-        digital_object = Curator::DigitalObject.create! valid_attributes
-
-        put :update, params: { id: digital_object.to_param, digital_object: invalid_attributes }, session: valid_session
-        expect(response).to have_http_status(:unprocessable_entity)
-        expect(response.content_type).to eq('application/json')
-      end
-    end
-  end
 end
